Explicitely support tmux and zellij, fallback if terminfo isn't available
Within tmux and zellij (and screen but that's pre-existing) we can assume the escape and avoid looking up terminfo. This can help on systems which are a bit bare or broken and don't have a functional terminfo. Even outside of these, we can detect for terminfo and not select the variant if we don't find one, which will fall back to the hardcoded sequence more often.